22.07.2020 Refactor code
