+++ requirements.txt
@@ -1,5 +1,6 @@
 certifi==2020.6.20
 chardet==3.0.4
+dj-database-url==0.5.0
 Django==2.2.14
 django-crispy-forms==1.9.2
 django-filter==2.3.0
